+++ grails-app/controllers/io/xh/hoist/impl/XhController.groovy
@@ -219,8 +219,8 @@ class XhController extends BaseController {
         renderJSON(environmentService.getEnvironment())
     }
 
-    def version() {
-        def options = configService.getMap('xhAppVersionCheck', [:])
+    def status() {
+        def options = configService.getMap('xhAppStatusCheck', [:])
         renderJSON(
             *: options,
             instanceName: clusterService.instanceName,