+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -99,16 +99,16 @@ class BootStrap implements LogSupport {
                 groupName: 'xh.io',
                 note: 'Official TimeZone for this application - e.g. the zone of the head office. Used to format/parse business related dates that need to be considered and displayed consistently at all locations. Set to a valid Java TimeZone ID.'
             ],
-            xhAppVersionCheck: [
+            xhAppStatusCheck: [
                 valueType: 'json',
                 defaultValue: [
-                    interval: configService.getInt('xhAppVersionCheckSecs', 30),
-                    mode: configService.getBool('xhAppVersionCheckEnabled', true) ? 'promptReload' : 'silent'
+                    interval: 10,
+                    updateMode: configService.getMap('xhAppVersionCheck', [mode: 'promptReload']).get('mode')
                 ],
                 clientVisible: true,
                 groupName: 'xh.io',
-                note: "Controls application behaviour when the server reports to the client that a new version is available. Supports the following options:\n\n" +
-                    "-'interval': Frequency (in seconds) with which the version of the app should be checked. Value of -1 disables version checking.\n" +
+                note: "Controls client calls to server to poll for new version and/or instance. Supports the following options:\n\n" +
+                    "-'interval': Frequency (in seconds) with which the status of the app server should be polled. Value of -1 disables checking.\n" +
                     "-'mode': Action taken by client upon a new version becoming available. Supports the following options:\n" +
                     "\t+ 'forceReload': Force clients to refresh immediately. To be used when an updated server is known to be incompatible with a previously deployed client.\n" +
                     "\t+ 'promptReload': Show an update prompt banner, allowing users to refresh when convenient.\n" +